refactor: add abstract base class for User
We needed a way to explicitly represent the {pre,post}-conditions
of the generated User object.
