+++ src/authn/user/User.js
@@ -1,0 +1,9 @@
+export class User {
+  /**
+   * @param {string} givenPassword
+   * @returns {Promise<boolean>}
+   */
+  async isValidPassword(givenPassword) {
+    throw new Error("Not Implemented");
+  }
+}
